requirements.txt
aws_cdk.core
aws_cdk.aws_events
aws_cdk.aws_events_targets
aws_cdk.aws_glue
aws_cdk.aws_iam
aws_cdk.aws_kinesisfirehose
aws_cdk.aws_lambda
aws_cdk.aws_lambda_python
aws_cdk.aws_logs
aws_cdk.aws_s3
aws_cdk.aws_s3_notifications
aws_cdk.aws_securityhub
aws_cdk.aws_sns
aws_cdk.aws_sns_subscriptions
aws_cdk.aws_sqs
aws_cdk.custom_resources
aws_cdk.aws_athena
aws_cdk.aws_quicksight
aws_cdk.aws_s3_deployment
boto3
